moved create account to pages instead of components
